revert BlockNumber to uint32 in lieu of better storage slot packing

+++ test/unit/CoreHash.t.sol
@@ -187,7 +187,7 @@ contract CoreHashTest is Base {
     function test_coreHash_fuzz(
         bytes32 key,
         address creator,
-        uint64 rawCreatedAt,
+        uint32 rawCreatedAt,
         string calldata contentType,
         bytes calldata payload
     ) public {
